requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==2.0.2
sqlalchemy==2.0.37
scikit-learn==1.6.1
python-binance==1.0.27
python-dotenv==1.0.1
PyYAML
dash==2.14.1
plotly==5.18.0
psycopg2-binary==2.9.9
matplotlib==3.9.0